==> 1.0 extract data from xml.ipynb <==
# import libraries
import xml.etree.ElementTree as et
import os
import pandas as pd
from xml.parsers import expat

# Define column names and folders
col_record_id = "Record ID"
col_title = "Title"
col_abstract = "Abstract"
col_publication_year= "Publication Year"
col_num_author = "Number of Authors"
col_geographic_term = "Geographic Term"
col_index_term = "Index Term"
col_publisher= "Publisher"
col_published_on = "Published On"
col_conference = "Conference"
col_conference_location = "Conference Location"
col_volume = "Volume"
col_issue = "Issue"
col_type = "Type"
col_language = "Language"
col_issn = "ISSN"
col_eissn = "EISSN"

all_papers_folder = 'xml_data/'

#Read xml files from defined folder and save as pkl format

path = all_papers_folder
folder = os.fsencode(path)
list_dic = []

for file in os.listdir(folder):
    filename = os.fsdecode(file)
    print(filename)
    #<?xml version='1.0' encoding='iso-8859-1'?>
    parser = et.XMLParser(encoding="iso-8859-1")
    tree = et.parse(path+filename, parser = parser)
    root = tree.getroot()
    
    for record in root:
        dictionary = {}
        if "id" in record.attrib:
            dictionary[col_record_id] = record.attrib["id"]
        #find the title
        title = record.find("title")
        if title!=None:
            dictionary[col_title] = title.text
            #print(title.text)
        # find language
        language = record.find("language_1")
        if language != None:
            dictionary[col_language] = language.text
        
        # find abstract
        abstract = record.find("abstract")
        if abstract!=None:
            dictionary[col_abstract] = abstract.text
            
        # find the Publication Year + publisher + published on
        document = record.find("document")
        if document!=None:
            authors = document.find("authors")
            if authors!=None:
                dictionary[col_num_author] = len(authors)

            monograph = document.find("monograph")
            if monograph!= None:
                
                volume = monograph.find("volume")
                if volume!=None:
                    dictionary[col_volume] = volume.text
                
                issue = monograph.find("issue")
                if issue!= None:
                    dictionary[col_issue] = issue.text
                
                publication_date = monograph.find("publication_date")
                if publication_date!= None:
                    dictionary[col_publication_year] =  int(publication_date.text)
                    #print(publication_date.text)

                serial = monograph.find("serial")
                if serial!=None:
                    #print(serial.attrib)
                    if "publisher" in serial.attrib:
                        publisher = serial.attrib["publisher"]
                    else:
                        publisher = ""
                        
                    if "issn" in serial.attrib:
                        dictionary[col_issn] = serial.attrib["issn"]
                    else:
                        dictionary[col_issn] = ""
                    
                    if "eissn" in serial.attrib:
                        dictionary[col_eissn] = serial.attrib["eissn"]
                    else:
                        dictionary[col_eissn] = ""
                        
                    published_on = serial.text
                    dictionary[col_publisher] = publisher
                    dictionary[col_published_on] = published_on

                conference = monograph.find("conference")
                conference_string = None
                if conference!=None:
                    conference_string = conference.text

                    conference_location_string = None
                    if "country" in conference.attrib:
                        conference_location_string = conference.attrib["country"]
                    
                    dictionary[col_conference_location] = conference_location_string
                dictionary[col_conference] = conference_string
                    

        index_terms = record.find("index_terms")
        geo_terms = []
        index_term_list = []
        if index_terms!= None:
            for term in index_terms:
                if "type" in term.attrib:
                    if term.attrib["type"] == "GT":
                        geo_terms.append(term.text)
                    elif term.attrib["type"] == "IT":
                        index_term_list.append(term.text)
            
            #print(geo_terms)
        dictionary[col_index_term] = index_term_list
        dictionary[col_geographic_term] = geo_terms
        
        subject_areas = record.find("subject_areas")
        sub_area_list = []
        if subject_areas != None:
            for area in subject_areas:
                sub_area_list.append(area.text)
                
        dictionary["Subject Area"] = sub_area_list
        list_dic.append(dictionary)
        
table = pd.DataFrame(list_dic)

table.head()

#Check if there is any duplicate document
sum(table.duplicated(subset=["Record ID"]))

# Categorize documents as journal, conference or both

def determine_paper_type(row):
    if pd.isnull(row[col_conference]):
        if pd.isnull(row[col_issue]) and pd.isnull(row[col_volume]):
            return ""
        else:
            return "Journal"
    else:
        if pd.isnull(row[col_issue]) and pd.isnull(row[col_volume]):
            return "Conference"
        else:
            return "Both"

table[col_type] = table.apply(determine_paper_type, axis=1)

table.count()

# save as pkl and excel

table.to_pickle("files/1.extracted data from xml.pkl")

#table[table[col_abstract].str.contains("This paper presents a gap-based solution", na=False)].iloc[0][col_abstract]

table = pd.read_pickle("files/1.extracted data from xml.pkl")

len(table)

", ".join(list(table.columns))
