==> 2.1 clean table removing items.ipynb <==
import pandas as pd
import en_core_web_sm
from langdetect import detect


table = pd.read_pickle("files/1.extracted data from xml.pkl")

col_title = "Title"
col_abstract = "Abstract"
col_publication_year= "Publication Year"
col_num_author = "Number of Authors"
col_geographic_term = "Geographic Term"
col_publisher= "Publisher"
col_published_on = "Published On"
col_conference = "Conference"
col_conference_location = "Conference Location"
col_volume = "Volume"
col_issue = "Issue"
col_type = "Type"
col_language = "Language"
col_abstract_clean = "Abstract Clean"
col_abstract_lemma = "Abstract Lemma"
all_papers_folder = 'files/'

# Remove rows where abstract is less than 100 character
print("Initial "+str(len(table)))
table.drop(table.index[table[col_abstract].str.len()<100], inplace = True)
print("After len < 100   "+str(len(table)))
table.drop(table.index[pd.isnull(table[col_abstract])], inplace = True)
print("After null abstract = "+str(len(table)))

table  = table[(pd.isnull(table[col_language])) | (table[col_language] == "English")]
print("After only English = "+str(len(table)))

valid_journals = pd.read_pickle("files/2.0 identified_sci_ssci_ei_journals.pkl")
invalid_journal_name_list = valid_journals[valid_journals["Selected"]==False]["Published On"].tolist()


table.drop(table.index[table[col_published_on].isin(invalid_journal_name_list)], inplace = True)

print("After deleting invalid journals = "+str(len(table)))

# Remove rows where publication type is nothing
table.drop(table.index[table["Type"] ==""], inplace = True)
print("After colum type =="" = "+str(len(table)))

detected_lang = []
for i, r in table.iterrows():
    if i%10000 == 0:
        print(i)
    lang = detect(r[col_abstract_clean])
    detected_lang.append(lang)
        
table['detected_lang'] = detected_lang

from collections import Counter
print(Counter(detected_lang))

table[table.detected_lang!="en"]

table[(table.detected_lang!="en") & (table.Language=="English")]

table = table[table.detected_lang=="en"]

table.to_pickle("files/2.1 selected row data.pkl")
table.to_csv("files/2.1 selected row data.csv")
