==> 3.0 clean_text_data.ipynb <==
import pandas as pd
import spacy
# spaCy based imports
from nltk.corpus import stopwords
from spacy.lang.en.stop_words import STOP_WORDS
from spacy.lang.en import English
import en_core_web_sm
import string
import re
from langdetect import detect

# Download and install en_core_web_sm
# !python -m spacy download en_core_web_sm

table = pd.read_pickle("files/2.1 selected row data.pkl")
table = table.reset_index(drop=True)
col_title = "Title"
col_abstract = "Abstract"
col_publication_year= "Publication Year"
col_num_author = "Number of Authors"
col_geographic_term = "Geographic Term"
col_publisher= "Publisher"
col_published_on = "Published On"
col_conference = "Conference"
col_conference_location = "Conference Location"
col_volume = "Volume"
col_issue = "Issue"
col_type = "Type"
col_abstract_clean = "Abstract Clean"
col_abstract_lemma = "Abstract Lemma"
col_geo_clean = col_geographic_term+"clean"
col_geo_sanitized = "Geographic term Sanitized"
all_papers_folder = 'files/'

table

# Creating a spaCy object
nlp = en_core_web_sm.load()

# list of allowed parts of speech
allowed_postags=['NOUN', 'ADJ', 'VERB', 'ADV', 'PROPN']

#Lemmatization (or less commonly lemmatization) in linguistics is the process of grouping together the inflected forms of a word so they can be analysed as a single item, identified by the word's lemma, or dictionary form.
# For example, the word "walk" is the base form for the word "walking", and hence this is matched in both stemming and lemmatization

def lemmatize_with_spacy(sentence):
    doc = nlp(sentence)
    mytokens = [token.lemma_ for token in doc if token.pos_ in allowed_postags and token.lemma_ != "-PRON-"]
    mytokens = " ".join([i for i in mytokens])
    return mytokens

# converting special characters to standard
def clean_abstract(row):
    text = row[col_abstract]
    text = text.replace("\x91", "'")
    text = text.replace("\x92", "'")
    text = text.replace("\x93", '"')
    text = text.replace("\x94", '"')
    text = text.replace("\x96", '-')
    text = text.replace("\x97", '--')
    text = text.replace("%", ' ')
    text = re.sub(r"[^a-zA-Z -'""]"," ",text)
    
    for i in range(9):
        text= text.replace("  ", " ")
    
    return text

table[col_abstract_clean] = table.apply(clean_abstract, axis=1)

table.head()

lemma_list = []
for i, r in table.iterrows():
    if i%10000 ==0: 
        print(i)
    lemma_list.append(lemmatize_with_spacy(r[col_abstract_clean]))
table[col_abstract_lemma] = lemma_list

i = 0
print(table.iloc[i][col_abstract])
print()
print(table.iloc[i][col_abstract_clean])
print()
print(table.iloc[i][col_abstract_lemma])


table.sample(7)

# Clean Geographic information

import pycountry

states = set(['Alabama','Alaska','Arizona','Arkansas','California','Colorado','Connecticut','Delaware', 'Florida','Georgia','Hawaii','Idaho','Illinois','Indiana','Iowa','Kansas','Kentucky','Louisiana','Maine','Maryland','Massachusetts','Michigan','Minnesota','Mississippi','Missouri','Montana','Nebraska','Nevada','New Hampshire','New Jersey','New Mexico','New York','North Carolina','North Dakota','Ohio','Oklahoma','Oregon','Pennsylvania','Rhode Island','South Carolina','South Dakota','Tennessee','Texas','Utah','Vermont','Virginia','Washington','West Virginia','Wisconsin','Wyoming', "District of Columbia"])
print(len(states))
countries = set([country.name for country in pycountry.countries])
countries = countries.union(set(['Bolivia', 'Korea', 'Palestine', 'Russia', 'Taiwan', 'Tanzania', 'Vietnam', 'Iran', "Scotland", "England", "Czech Republic"]))
len(countries)

def detect_geographic_region(row):
    if len(row[col_geographic_term]) == 0 and not pd.isnull(row[col_abstract]):
        doc = nlp(row[col_abstract_clean])
        locations = set()
        for x in doc.ents:
            if x.label_ == "GPE":
                locations.add(x.text)
        return list(locations)
    else:
        return row[col_geographic_term]

table[col_geo_clean] = table.apply(detect_geographic_region, axis=1)

def sanitize_country_name(row):
    saved_list = row[col_geo_clean]
    new_list = []
    for c in countries:
        new_list.extend([c for saved_country in saved_list if c.lower() in saved_country.lower()])
    
    for s in states:
        new_list.extend(["United States" for saved_country in saved_list if s.lower() in saved_country.lower()])
    
    if "England" in new_list:
        new_list.remove("England")
        new_list.append("United Kingdom")
    
    c_set = set(new_list)
    
    if "Oman" in c_set and "Romania" in c_set:
        c_set.remove("Oman")
    
    if "United States" in c_set and "Jersey" in c_set:
        c_set.remove("Jersey")
    return list(c_set)

def sanitize_state_name(row):
    saved_list = row[col_geo_clean]
    new_list = []
    if "United States" in row["Sanitize"]:
        for s in states:
            new_list.extend([s for saved_country in saved_list if s.lower() in saved_country.lower()])
    
    c_set = set(new_list)
    
    if "Washington" in c_set and "District of Columbia" in c_set:
        c_set.remove("Washington")
    
    return list(c_set)

table["Sanitize"] = table.apply(sanitize_country_name, axis=1)
table["Sanitize State"] = table.apply(sanitize_state_name, axis=1)

table.to_pickle("files/3.0 data_clean_text_region.parquet")
table.to_csv("files/3.0 data_clean_text_region.csv")
