==> 4.1 run_lda_save_model.ipynb <==
%%time
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.datasets import fetch_20newsgroups
from sklearn.decomposition import NMF, LatentDirichletAllocation
import pandas as pd
from nltk.corpus import stopwords 
import pickle
import numpy as np

#Download stopwords in not downloaded
#nltk.download('stopwords')

col_title = "Title"
col_abstract = "Abstract"
col_abstract_clean = "Abstract Clean"
col_abstract_lemma = "Abstract Lemma"
col_publication_year= "Publication Year"
col_num_author = "Number of Authors"
col_geographic_term = "Geographic Term"
col_publisher= "Publisher"
col_published_on = "Published On"
col_conference = "Conference"
col_conference_location = "Conference Location"
col_volume = "Volume"
col_issue = "Issue"
col_type = "Type"

stop_words    = set(stopwords.words('english'))
print(len(stop_words))
extra_stopwords = ['use','model', 'result','study','system','vehicle','-pron-', '-PRON-', 'paper', 'author', 'method', 'propose', 'datum','have','that','can','which','datum','der','author','also','more','not','und','such','other','when','may','will','however','there','article','how','only','von','then','could','den','non','where','further','eine','auf','das','des','ein','bei','who','thus','ist','often','zur','einer','must','able','aus','zum','pre','einem',]
stop_words = stop_words.union(set(extra_stopwords))
len(stop_words)
#extra_stopwords

table = pd.read_pickle("files/3.0 data_clean_text.pkl")
documents = table[col_abstract_lemma].tolist()

# import nltk
# nltk.download('punkt')

# find the most frequent words and look for something that is not relevant to determining topics and add them to stopword list
# all_words = word_tokenize(" ".join(documents))
# print("Total number of words =", len(all_words))
# word_count = Counter(all_words)
# df = pd.DataFrame({"word":list(word_count.keys()), "counts":list(word_count.values())})
# df = df.sort_values(by="counts", ascending=False)
# df.head(10)

%%time
# LDA can only use raw term counts for LDA because it is a probabilistic graphical model
#https://scikit-learn.org/stable/modules/generated/sklearn.feature_extraction.text.CountVectorizer.html

# max_features = top n number of words to keep
# max_df = if 100: removes words which are in more than 100 documents_____ if .95: removes word which are in more than 95% of the documents
# min_df = if 5: removes words which are in less than 5 documents
# stop_words = 'english' removes stopwords in english
# ngram_range=(1,2): if my text contains the phrase "butter chicken" in higer number of times: it will be considered like a word


tf_vectorizer = CountVectorizer(max_df=.6, min_df=.005, stop_words=list(stop_words))
tf = tf_vectorizer.fit_transform(documents)
tf_feature_names = tf_vectorizer.get_feature_names_out()

x_traincv_df = pd.DataFrame(tf.toarray(),columns=list(tf_feature_names))
x_traincv_df.to_parquet("files/4_document_word_matrix.parquet")

%%time
# Run LDA. Topic is set to 50 based on the literature and judgement regarding transportation fields
no_topics = 50
lda = LatentDirichletAllocation(n_components=no_topics, max_iter=10, learning_method='batch', verbose=2,n_jobs=-1).fit(tf)


with open('files/4.0 lda_model.pickle', 'wb') as f:
    pickle.dump( lda, f)

# Functions for printing keywords for each topic
def get_topic_word_d_in_map(model, vectorizer, top_n=10):
    d ={}
    for idx, topic in enumerate(model.components_):
        print("Topic %d:" % (idx))
        # print([(vectorizer.get_feature_names()[i], topic[i]) for i in topic.argsort()[:-top_n - 1:-1]]) 
        # print([(vectorizer.get_feature_names()[i], topic[i]) for i in topic.argsort()[:-top_n - 1:-1]]) 
        d[idx] = [(vectorizer.get_feature_names_out()[i], topic[i]) for i in topic.argsort()[:-top_n - 1:-1]]
    return d

#Save data for later use
# map key = topic number; value = (word, weight)

get_topic_word_d_in_map = get_topic_word_d_in_map(lda, tf_vectorizer, 100)
with open('files/4.1 topic_word_weight.pickle', 'wb') as f:
    pickle.dump( get_topic_word_d_in_map, f)

#d = pickle.load(open("4_topic_word_weight.pickle", 'rb'))

get_topic_word_d_in_map[0]

# doc_topic indicates the probability of corresponding to the topic(columns)  for each document (row)
doc_topic = lda.transform(tf)
doc_topic_most_probable=[]

for n in range(doc_topic.shape[0]):
    topic_most_pr = doc_topic[n].argmax()
    # print("doc: {} topic: {}\n".format(n,topic_most_pr))
    doc_topic_most_probable.append(topic_most_pr)


doc_topic.shape

doc_topic_most_probable[:10]

print(doc_topic[0])
doc_topic_most_probable[0]


print(len(doc_topic))
print(len(doc_topic_most_probable))

table["Topic Most Probable"] = doc_topic_most_probable

table.head()

doc_topic_all_topic=[]
for n in range(doc_topic.shape[0]):
    doc_topic_all_topic.append(doc_topic[n])
table["Topic All"] = doc_topic_all_topic

table.head()

# some topics are related to group of words which are a cluster of academic words we replace the topic of those document with the second most probable topic
def assignSecondTopic(row):
    topic=row["Topic Most Probable"]
    if row["Topic Most Probable"] in [27, 34, 49]:
        secondMost = np.where(row["Topic All"]==np.sort(row["Topic All"])[-2])
        topic =  secondMost
    return topic

table["Topic Most Probable"]  = table.apply(assignSecondTopic, axis=1)

table.to_parquet("files/4.0 data_with_topic.parquet")
table.to_csv("files/4.0 data_with_topic.csv")
